# Use WithLabelValues instead of With for metrics in atime_updater