This should save a bit of CPU in the cache proxy.

Benchmark results:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor
           │    base     │               values               │
           │   sec/op    │   sec/op     vs base               │
Enqueue-24   146.6m ± 4%   138.0m ± 3%  -5.86% (p=0.005 n=11)

           │     base     │                values                │
           │     B/op     │     B/op      vs base                │
Enqueue-24   689.3Mi ± 0%   400.9Mi ± 0%  -41.84% (p=0.000 n=11)

           │    base     │               values                │
           │  allocs/op  │  allocs/op   vs base                │
Enqueue-24   7.531M ± 0%   5.730M ± 0%  -23.90% (p=0.000 n=11)
```